refactor(estructure): interface, service and DTO factorization
