+++ src/Bio.Backend.Core/Bio.API/Controllers/UserRolesController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Bio.Application.DTOs;
 using Bio.Application.Services;
+using Bio.Application.Interfaces;
 
 namespace Bio.API.Controllers;
 
@@ -25,7 +26,7 @@ public class UserRolesController : ControllerBase
     /// <returns>A list of assignments.</returns>
     /// <response code="200">List of assignments retrieved.</response>
     [HttpGet]
-    [ProducesResponseType(typeof(IEnumerable<UserRoleReadDTO>), StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(IEnumerable<UserRoleResponseDTO>), StatusCodes.Status200OK)]
     public async Task<IActionResult> GetAssignments()
     {
         var assignments = await _userRoleService.GetAllAssignmentsAsync();
@@ -38,7 +39,7 @@ public class UserRolesController : ControllerBase
     /// <param name="userId">The unique identifier of the user.</param>
     /// <returns>A list of roles assigned to that user.</returns>
     [HttpGet("user/{userId:guid}")]
-    [ProducesResponseType(typeof(IEnumerable<UserRoleReadDTO>), StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(IEnumerable<UserRoleResponseDTO>), StatusCodes.Status200OK)]
     public async Task<IActionResult> GetByUser(Guid userId)
     {
         var assignments = await _userRoleService.GetAssignmentsByUserIdAsync(userId);
@@ -51,7 +52,7 @@ public class UserRolesController : ControllerBase
     /// <param name="roleName">The name of the role.</param>
     /// <returns>A list of user-role assignments for that role.</returns>
     [HttpGet("role/{roleName}")]
-    [ProducesResponseType(typeof(IEnumerable<UserRoleReadDTO>), StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(IEnumerable<UserRoleResponseDTO>), StatusCodes.Status200OK)]
     public async Task<IActionResult> GetByRole(string roleName)
     {
         var assignments = await _userRoleService.GetAssignmentsByRoleNameAsync(roleName);
@@ -64,7 +65,7 @@ public class UserRolesController : ControllerBase
     /// <param name="roleId">The unique identifier of the role.</param>
     /// <returns>A list of user-role assignments for that role ID.</returns>
     [HttpGet("role-id/{roleId:guid}")]
-    [ProducesResponseType(typeof(IEnumerable<UserRoleReadDTO>), StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(IEnumerable<UserRoleResponseDTO>), StatusCodes.Status200OK)]
     public async Task<IActionResult> GetByRoleId(Guid roleId)
     {
         var assignments = await _userRoleService.GetAssignmentsByRoleIdAsync(roleId);
