+++ src/Bio.Backend.Core/Bio.Application/DTOs/RoleResponseDTO.cs
@@ -5,9 +5,28 @@ namespace Bio.Application.DTOs;
 /// </summary>
 public class RoleResponseDTO
 {
+    /// <summary>
+    /// Unique identifier for the role.
+    /// </summary>
     public Guid Id { get; set; }
+
+    /// <summary>
+    /// Name of the role.
+    /// </summary>
     public string Name { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Description of the role.
+    /// </summary>
     public string? Description { get; set; }
+
+    /// <summary>
+    /// Timestamp of when the role was created.
+    /// </summary>
     public DateTime CreatedAt { get; set; }
+
+    /// <summary>
+    /// Timestamp of when the role was last updated.
+    /// </summary>
     public DateTime? UpdatedAt { get; set; }
 }
