+++ src/Bio.Backend.Core/Bio.Application/DTOs/RoleUpdateDTO.cs
@@ -9,10 +9,13 @@ public class RoleUpdateDTO
     /// Gets or sets the name of the role.
     /// Case-insensitive.
     /// </summary>
+    [Required(ErrorMessage = "Name is required.")]
+    [StringLength(100, ErrorMessage = "Name cannot exceed 100 characters.")]
     public string Name { get; set; } = string.Empty;
 
     /// <summary>
     /// Gets or sets the optional description of the role.
     /// </summary>
+    [StringLength(2000, ErrorMessage = "Description cannot exceed 2000 characters.")]
     public string? Description { get; set; }
 }
