+++ src/Bio.Backend.Core/Bio.Application/DTOs/UserRoleCreateDTO.cs
@@ -10,12 +10,14 @@ public class UserRoleCreateDTO
     /// <summary>
     /// The unique identifier of the user.
     /// </summary>
-    [Required]
+    /// <example>123e4567-e89b-12d3-a456-426614174000</example>
+    [Required(ErrorMessage = "User ID is required.")]
     public Guid UserId { get; set; }
 
     /// <summary>
     /// The unique identifier of the role to assign.
     /// </summary>
-    [Required]
+    /// <example>123e4567-e89b-12d3-a456-426614174000</example>
+    [Required(ErrorMessage = "Role ID is required.")]
     public Guid RoleId { get; set; }
 }
