+++ src/Bio.Backend.Core/Bio.UnitTests/Application/Services/UserRoleServiceTests.cs
@@ -1,5 +1,6 @@
 using Bio.Application.DTOs;
 using Bio.Application.Services;
+using Bio.Application.Interfaces;
 using Bio.Domain.Entities;
 using Bio.Domain.Interfaces;
 using FluentAssertions;
